Add range slider support (#1275)

+++ src/lv_misc/lv_fs.c
@@ -539,7 +539,7 @@ char * lv_fs_get_letters(char * buf)
  */
 const char * lv_fs_get_ext(const char * fn)
 {
-    size_t i;
+    uint16_t i;
     for(i = strlen(fn); i > 0; i--) {
         if(fn[i] == '.') {
             return &fn[i + 1];
@@ -558,7 +558,7 @@ const char * lv_fs_get_ext(const char * fn)
  */
 char * lv_fs_up(char * path)
 {
-    size_t len = strlen(path);
+    uint16_t len = strlen(path);
     if(len == 0) return path;
 
     len--; /*Go before the trailing '\0'*/
@@ -589,7 +589,7 @@ char * lv_fs_up(char * path)
  */
 const char * lv_fs_get_last(const char * path)
 {
-    size_t len = strlen(path);
+    uint16_t len = strlen(path);
     if(len == 0) return path;
 
     len--; /*Go before the trailing '\0'*/
